# tests/contracts/openapi.public.yaml
openapi: 3.0.3
info:
  title: Nexus EAF Public API Contracts
  version: 1.0.0
servers:
  - url: http://127.0.0.1:3000
paths:
  /api/v1/health:
    get:
      summary: Health endpoint
      responses:
        '200':
          description: Service healthy
          content:
            application/json:
              schema:
                type: object
                required: [status, checks, timestamp, release, voice]
                properties:
                  status: { type: string }
                  checks:
                    type: object
                    required: [db, app]
                    properties:
                      db: { type: string }
                      app: { type: string }
                  timestamp: { type: string, format: date-time }
                  release:
                    type: object
                    required: [gitSha, buildTime, nodeEnv]
                    properties:
                      gitSha: { type: string }
                      buildTime: { type: string }
                      nodeEnv: { type: string }
                  voice:
                    type: object
                    required: [requestedVoiceMode, effectiveVoiceMode, sttAvailable, ttsAvailable]
                    properties:
                      requestedVoiceMode: { type: string }
                      effectiveVoiceMode: { type: string }
                      sttAvailable: { type: boolean }
                      ttsAvailable: { type: boolean }
        '503':
          description: Service degraded
          content:
            application/json:
              schema:
                type: object
                required: [status, checks, timestamp, release, voice]
                properties:
                  status: { type: string }
                  checks:
                    type: object
                    required: [db, app]
                    properties:
                      db: { type: string }
                      app: { type: string }
                  timestamp: { type: string, format: date-time }
                  release:
                    type: object
                    required: [gitSha, buildTime, nodeEnv]
                    properties:
                      gitSha: { type: string }
                      buildTime: { type: string }
                      nodeEnv: { type: string }
                  voice:
                    type: object
                    required: [requestedVoiceMode, effectiveVoiceMode, sttAvailable, ttsAvailable]
                    properties:
                      requestedVoiceMode: { type: string }
                      effectiveVoiceMode: { type: string }
                      sttAvailable: { type: boolean }
                      ttsAvailable: { type: boolean }

  /api/v1/auth/register:
    post:
      summary: Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                # Major / non mineur
                - type: object
                  additionalProperties: false
                  required: [email, password, acceptedCgu]
                  properties:
                    email:
                      type: string
                      maxLength: 254
                      pattern: "^[A-Za-z0-9][A-Za-z0-9_+\\-]{0,63}@(?:[A-Za-z0-9](?:[A-Za-z0-9\\-]{0,61}[A-Za-z0-9])?\\.)+[A-Za-z]{2,}$"
                    password:
                      type: string
                      minLength: 8
                      maxLength: 128
                      allOf:
                        - pattern: ".*[a-z].*"
                        - pattern: ".*[A-Z].*"
                        - pattern: ".*[0-9].*"
                    displayName:
                      type: string
                      minLength: 1
                      maxLength: 120
                    acceptedCgu:
                      type: boolean
                      enum: [true]
                    cguVersion:
                      type: string
                      minLength: 1
                      maxLength: 20
                      pattern: "^[A-Za-z0-9._\\-]{1,20}$"
                    isMinor:
                      type: boolean
                      enum: [false]
                # Mineur
                - type: object
                  additionalProperties: false
                  required: [email, password, acceptedCgu, isMinor, parentEmail]
                  properties:
                    email:
                      type: string
                      maxLength: 254
                      pattern: "^[A-Za-z0-9][A-Za-z0-9_+\\-]{0,63}@(?:[A-Za-z0-9](?:[A-Za-z0-9\\-]{0,61}[A-Za-z0-9])?\\.)+[A-Za-z]{2,}$"
                    password:
                      type: string
                      minLength: 8
                      maxLength: 128
                      allOf:
                        - pattern: ".*[a-z].*"
                        - pattern: ".*[A-Z].*"
                        - pattern: ".*[0-9].*"
                    displayName:
                      type: string
                      minLength: 1
                      maxLength: 120
                    acceptedCgu:
                      type: boolean
                      enum: [true]
                    cguVersion:
                      type: string
                      minLength: 1
                      maxLength: 20
                      pattern: "^[A-Za-z0-9._\\-]{1,20}$"
                    isMinor:
                      type: boolean
                      enum: [true]
                    parentEmail:
                      type: string
                      maxLength: 254
                      pattern: "^[A-Za-z0-9][A-Za-z0-9._+\\-]{0,63}@(?:[A-Za-z0-9](?:[A-Za-z0-9\\-]{0,61}[A-Za-z0-9])?\\.)+[A-Za-z]{2,}$"
      responses:
        '201':
          description: Registered
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok: { type: boolean }
        '400':
          description: Invalid payload
          content:
            application/json:
              schema:
                type: object
                required: [error]
                properties:
                  error: { type: string }
        '409':
          description: Existing account
          content:
            application/json:
              schema:
                type: object
                required: [error]
                properties:
                  error: { type: string }
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                type: object
                required: [error]
                properties:
                  error: { type: string }

  /api/v1/auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, password]
              properties:
                email:
                  type: string
                  maxLength: 254
                  pattern: "^[A-Za-z0-9][A-Za-z0-9_+\\-]{0,63}@(?:[A-Za-z0-9](?:[A-Za-z0-9\\-]{0,61}[A-Za-z0-9])?\\.)+[A-Za-z]{2,}$"
                password:
                  type: string
                  minLength: 1
      responses:
        '200':
          description: Login success
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok: { type: boolean }
        '400':
          description: Invalid payload
          content:
            application/json:
              schema:
                type: object
                required: [error]
                properties:
                  error: { type: string }
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: object
                required: [error]
                properties:
                  error: { type: string }
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                type: object
                required: [error]
                properties:
                  error: { type: string }